+++ src/play/modules/elasticsearch/mapping/impl/SimpleFieldMapper.java
@@ -5,8 +5,6 @@ import java.lang.reflect.Field;
 
 import org.elasticsearch.common.xcontent.XContentBuilder;
 
-import play.modules.elasticsearch.util.ReflectionUtil;
-
 /**
  * Field mapper for simple, single-valued types
  * 
@@ -21,7 +19,7 @@ public class SimpleFieldMapper<M> extends AbstractFieldMapper<M> {
 
 	@Override
 	public void addToMapping(XContentBuilder builder, String prefix) throws IOException {
-		String name = field.getName();
+		String name = getFieldName();
 		String type = getFieldType();
 
 		if (prefix != null) {
@@ -33,8 +31,8 @@ public class SimpleFieldMapper<M> extends AbstractFieldMapper<M> {
 
 	@Override
 	public void addToDocument(M model, XContentBuilder builder, String prefix) throws IOException {
-		String name = field.getName();
-		Object value = ReflectionUtil.getFieldValue(model, field);
+		String name = getFieldName();
+		Object value = getFieldValue(model);
 
 		if (value != null) {
 			if (prefix != null) {
@@ -42,17 +40,6 @@ public class SimpleFieldMapper<M> extends AbstractFieldMapper<M> {
 			} else {
 				builder.field(name, value);
 			}
-		}
-	}
-
-	protected String getFieldType() {
-		if (meta != null && meta.type().length() > 0) {
-			// Type was explicitly set, use it
-			return meta.type();
-
-		} else {
-			// Detect type automatically
-			return detectFieldType(field.getType());
 		}
 	}
